+++ src/module.js
@@ -1,5 +1,3 @@
-(function() {
-  angular.module('validation', ['validation.provider', 'validation.directive']);
-  angular.module('validation.provider', []);
-  angular.module('validation.directive', ['validation.provider']);
-}).call(this);
+angular.module('validation', ['validation.provider', 'validation.directive']);
+angular.module('validation.provider', []);
+angular.module('validation.directive', ['validation.provider']);