+++ test/unit/validationGroupSpec.js
@@ -58,6 +58,21 @@ describe('validation-group directive', function() {
       expect(element.hasClass('ng-invalid')).toBe(true);
     });
 
+    it('should be valid when at least one of elements is valid', function() {
+      $scope.Form.checkbox1.$setViewValue(true);
+
+      expect($scope.Form.$valid).toBe(true);
+      expect(element.hasClass('ng-valid')).toBe(true);
+
+      $scope.Form.checkbox1.$setViewValue(false);
+      expect($scope.Form.$valid).toBe(false);
+      expect(element.hasClass('ng-invalid')).toBe(true);
+
+      $scope.Form.checkbox2.$setViewValue(true);
+      expect($scope.Form.$valid).toBe(true);
+      expect(element.hasClass('ng-valid')).toBe(true);
+    });
+
     it('should have a success message inside the #checkbox element when an element is valid', function() {
       $scope.Form.checkbox1.$setViewValue(true);
 
@@ -109,7 +124,7 @@ describe('validation-group directive', function() {
       $compile = $injector.get('$compile');
       $scope = $rootScope.$new();
 
-      element = $compile('<form name="Form"><input type="text" name="email" ng-model="email" validator="required" validation-group="contact"><input type="number" name="telephone" ng-model="telephone" validator="number" validation-group="contact"><span id="contact"></span></form>')($scope);
+      element = $compile('<form name="Form"><input type="text" name="email" ng-model="email" validator="email" validation-group="contact"><input type="text" name="telephone" ng-model="telephone" validator="number" validation-group="contact"><span id="contact"></span></form>')($scope);
       angular.element(document.body).append(element);
       $scope.$digest();
     }));
@@ -139,14 +154,30 @@ describe('validation-group directive', function() {
     it('should be dirty and invalid', function() {
       $scope.Form.email.$setViewValue('foo@bar.com');
       $scope.Form.telephone.$setViewValue('065839481');
-      $scope.Form.email.$setViewValue();
-      $scope.Form.telephone.$setViewValue();
+      $scope.Form.email.$setViewValue('');
+      $scope.Form.telephone.$setViewValue('');
 
       expect($scope.Form.$dirty).toBe(true);
       expect(element.hasClass('ng-dirty')).toBe(true);
       expect($scope.Form.$invalid).toBe(true);
       expect(element.hasClass('ng-invalid')).toBe(true);
     });
+
+    it('should be valid when at least one of elements is valid', function() {
+      $scope.Form.email.$setViewValue('foo@bar.com');
+      expect($scope.Form.$valid).toBe(true);
+      expect(element.hasClass('ng-valid')).toBe(true);
+
+      $scope.Form.telephone.$setViewValue('065839481');
+      $scope.Form.email.$setViewValue('');
+      expect($scope.Form.$valid).toBe(true);
+      expect(element.hasClass('ng-valid')).toBe(true);
+
+      $scope.Form.telephone.$setViewValue('');
+      expect($scope.Form.$valid).toBe(false);
+      expect(element.hasClass('ng-invalid')).toBe(true);
+    });
+
 
     it('should have a success message inside the #contact element when an element is valid', function() {
       $scope.Form.email.$setViewValue('foo@bar.com');
@@ -157,7 +188,7 @@ describe('validation-group directive', function() {
 
     it('should have an error message inside the #contact element when no element is valid', function() {
       $scope.Form.email.$setViewValue('foo@bar.com');
-      $scope.Form.email.$setViewValue();
+      $scope.Form.email.$setViewValue('');
 
       messageElem = angular.element(element[0].querySelector('#contact > p'));
       expect(messageElem.hasClass('validation-invalid')).toBe(true);
@@ -174,7 +205,7 @@ describe('validation-group directive', function() {
     it('should have a success message inside the #contact element when one of element is valid', function() {
       $scope.Form.email.$setViewValue('foo@bar.com');
       $scope.Form.telephone.$setViewValue('065839481');
-      $scope.Form.email.$setViewValue();
+      $scope.Form.email.$setViewValue('');
 
       messageElem = angular.element(element[0].querySelector('#contact > p'));
       expect(messageElem.hasClass('validation-valid')).toBe(true);
@@ -183,8 +214,8 @@ describe('validation-group directive', function() {
     it('should have an error message inside the #contact element when both of elements are invalid', function() {
       $scope.Form.email.$setViewValue('foo@bar.com');
       $scope.Form.telephone.$setViewValue('065839481');
-      $scope.Form.email.$setViewValue();
-      $scope.Form.telephone.$setViewValue();
+      $scope.Form.email.$setViewValue('');
+      $scope.Form.telephone.$setViewValue('');
 
       messageElem = angular.element(element[0].querySelector('#contact > p'));
       expect(messageElem.hasClass('validation-invalid')).toBe(true);
@@ -230,6 +261,45 @@ describe('validation-group directive', function() {
       expect(errorSpy).not.toHaveBeenCalled();
     });
 
+    it('should validate a form and call a success callback when at least one of elements in the form is valid', function() {
+      successSpy = jasmine.createSpy('successSpy');
+      errorSpy = jasmine.createSpy('errorSpy');
+
+      $scope.Form.checkbox1.$setViewValue(true);
+
+      validationProvider.validate($scope.Form)
+        .success(function() {
+          successSpy();
+        })
+        .error(function() {
+          errorSpy();
+        });
+      $timeout.flush();
+      expect(successSpy).toHaveBeenCalled();
+      expect(errorSpy).not.toHaveBeenCalled();
+    });
+
+    it('should validate a form and call an error callback when all elements are invalid', function() {
+      successSpy = jasmine.createSpy('successSpy');
+      errorSpy = jasmine.createSpy('errorSpy');
+
+      $scope.Form.checkbox1.$setViewValue(true);
+      $scope.Form.checkbox2.$setViewValue(true);
+      $scope.Form.checkbox1.$setViewValue(false);
+      $scope.Form.checkbox2.$setViewValue(false);
+
+      validationProvider.validate($scope.Form)
+        .success(function() {
+          successSpy();
+        })
+        .error(function() {
+          errorSpy();
+        });
+      $timeout.flush();
+      expect(successSpy).not.toHaveBeenCalled();
+      expect(errorSpy).toHaveBeenCalled();
+    });
+
     it('should validate a form element and call a success callback', function() {
       successSpy = jasmine.createSpy('successSpy');
       errorSpy = jasmine.createSpy('errorSpy');
@@ -253,6 +323,7 @@ describe('validation-group directive', function() {
       errorSpy = jasmine.createSpy('errorSpy');
 
       $scope.Form.checkbox1.$setViewValue(true);
+      $scope.Form.checkbox1.$setViewValue(false);
 
       validationProvider.validate($scope.Form)
         .success(function() {
@@ -264,6 +335,76 @@ describe('validation-group directive', function() {
       $timeout.flush();
       expect(successSpy).not.toHaveBeenCalled();
       expect(errorSpy).toHaveBeenCalled();
+    });
+  });
+
+  describe('validation-group attribute validated by using the provider with expression dynamic return messages using functions. ', function() {
+    var successSpy;
+    var errorSpy;
+    beforeEach(inject(function($injector) {
+      $rootScope = $injector.get('$rootScope');
+      $compile = $injector.get('$compile');
+      validationProvider = $injector.get('$validation');
+      var expression = {
+        kuaisheng: function(value, scope, element, attrs, param) {
+          var errorStr = [
+            'errorStr1',
+            'errorStr2'
+          ];
+          var len = errorStr.length;
+          for (var i = len - 1; i >= 0; i--) {
+            if (value.indexOf(errorStr[i]) > -1) {
+              return {
+                result: false,
+                message: 'input should not include ' + errorStr[i]
+              };
+            }
+          }
+
+          return {
+            result: true,
+            message: ''
+          };
+        }
+      };
+      var defaultMsg = {
+        kuaisheng: {
+          error: 'valid is error'
+        }
+      };
+      validationProvider.setExpression(expression).setDefaultMsg(defaultMsg);
+
+      $timeout = $injector.get('$timeout');
+      $scope = $rootScope.$new();
+
+      element = $compile('<form name="Form"><input type="text" name="kuaisheng" ng-model="kuaisheng" validator="kuaisheng"></form>')($scope);
+      angular.element(document.body).append(element);
+      $scope.$digest();
+    }));
+
+    afterEach(function() {
+      element.remove();
+      element = null;
+    });
+
+    it('should validate a form and error message be dynamic', function() {
+      successSpy = jasmine.createSpy('successSpy');
+      errorSpy = jasmine.createSpy('errorSpy');
+
+      $scope.Form.kuaisheng.$setViewValue('asberrorStr1dd');
+
+
+      validationProvider.validate($scope.Form)
+        .success(function() {
+          successSpy();
+        })
+        .error(function() {
+          errorSpy();
+        });
+      $timeout.flush();
+      var messageElem = angular.element(element[0].querySelector('p'));
+      expect(messageElem.html()).toEqual('input should not include errorStr1');
+
     });
   });
 });