+++ webpack.config.js
@@ -1,0 +1,16 @@
+module.exports = {
+  entry: ['./src/module.js'],
+  output: {
+    filename: 'dist/bundle.js'
+  },
+  module: {
+    loaders: [{
+      test: /\.jsx?$/,
+      exclude: /(node_modules|bower_components)/,
+      loader: 'babel',
+      query: {
+        presets: ['es2015']
+      }
+    }]
+  }
+};